+++ src/Taxes/Contracts/TaxEngineDriver.php
@@ -14,12 +14,11 @@ declare(strict_types=1);
 
 namespace Vanilo\Taxes\Contracts;
 
+use Konekt\Extend\Contracts\Registerable;
 use Vanilo\Contracts\Address;
 use Vanilo\Contracts\Billpayer;
 
-interface TaxEngineDriver
+interface TaxEngineDriver extends Registerable
 {
     public function resolveTaxRate(Taxable $taxable, ?Billpayer $billpayer = null, ?Address $shippingAddress = null): ?TaxRate;
-
-    /** @todo Add this in v6 by extending `Registerable`: "public static function getName(): string;" */
 }