+++ src/Taxes/Tests/Dummies/DummyTaxDriver.php
@@ -24,6 +24,11 @@ class DummyTaxDriver implements TaxEngineDriver
 {
     public const TEST_RATE = 66;
 
+    public static function getName(): string
+    {
+        return 'dummy';
+    }
+
     public function resolveTaxRate(Taxable $taxable, ?Billpayer $billpayer = null, ?Address $shippingAddress = null): ?TaxRate
     {
         return new \Vanilo\Taxes\Models\TaxRate(['rate' => self::TEST_RATE]);